Fixed equality test.
Found a bug in equality test for scalar array fields. I wasn't atking
an absolute value of the difference, so if the difference was negative
it was always true. This revealed a bug in the implementation of the
derivative propagation for acosh and some bugs in the tests of the
derivative.